Split reader into stateful and stateless
