+++ dds/src/rtps/mod.rs
@@ -1,7 +1,6 @@
 pub mod behavior_types;
 pub mod cache_change;
 pub mod discovery_types;
-pub mod endpoint;
 pub mod entity;
 pub mod error;
 pub mod group;
@@ -9,10 +8,11 @@ pub mod message_receiver;
 pub mod message_sender;
 pub mod messages;
 pub mod participant;
-pub mod reader;
 pub mod reader_locator;
 pub mod reader_proxy;
+pub mod stateful_reader;
 pub mod stateful_writer;
+pub mod stateless_reader;
 pub mod stateless_writer;
 pub mod transport;
 pub mod types;
