+++ dds/src/rtps/stateful_reader.rs
@@ -1,5 +1,4 @@
 use super::{
-    endpoint::RtpsEndpoint,
     message_sender::MessageSender,
     messages::{
         self,
@@ -12,159 +11,12 @@ use super::{
     writer_proxy::RtpsWriterProxy,
 };
 use crate::{
-    implementation::{
-        data_representation_builtin_endpoints::{
-            discovered_reader_data::ReaderProxy, discovered_writer_data::WriterProxy,
-        },
-        data_representation_inline_qos::{
-            parameter_id_values::{PID_KEY_HASH, PID_STATUS_INFO},
-            types::{
-                StatusInfo, STATUS_INFO_DISPOSED, STATUS_INFO_DISPOSED_UNREGISTERED,
-                STATUS_INFO_FILTERED, STATUS_INFO_UNREGISTERED,
-            },
-        },
+    implementation::data_representation_builtin_endpoints::{
+        discovered_reader_data::ReaderProxy, discovered_writer_data::WriterProxy,
     },
-    transport::{
-        cache_change::CacheChange,
-        reader::ReaderHistoryCache,
-        types::{ChangeKind, ReliabilityKind},
-    },
+    transport::{cache_change::CacheChange, reader::ReaderHistoryCache, types::ReliabilityKind},
 };
 use tracing::error;
-
-impl CacheChange {
-    pub fn try_from_data_submessage(
-        data_submessage: &DataSubmessage,
-        source_guid_prefix: GuidPrefix,
-        source_timestamp: Option<messages::types::Time>,
-    ) -> Result<Self, String> {
-        let kind = match data_submessage
-            .inline_qos()
-            .parameter()
-            .iter()
-            .find(|&x| x.parameter_id() == PID_STATUS_INFO)
-        {
-            Some(p) => {
-                if p.length() == 4 {
-                    let status_info =
-                        StatusInfo([p.value()[0], p.value()[1], p.value()[2], p.value()[3]]);
-                    match status_info {
-                        STATUS_INFO_DISPOSED => Ok(ChangeKind::NotAliveDisposed),
-                        STATUS_INFO_UNREGISTERED => Ok(ChangeKind::NotAliveUnregistered),
-                        STATUS_INFO_DISPOSED_UNREGISTERED => {
-                            Ok(ChangeKind::NotAliveDisposedUnregistered)
-                        }
-                        STATUS_INFO_FILTERED => Ok(ChangeKind::AliveFiltered),
-                        _ => Err(format!(
-                            "Received invalid status info parameter with value {:?}",
-                            status_info
-                        )),
-                    }
-                } else {
-                    Err(format!(
-                        "Received invalid status info parameter length. Expected 4, got {:?}",
-                        p.length()
-                    ))
-                }
-            }
-            None => Ok(ChangeKind::Alive),
-        }?;
-
-        let instance_handle = match data_submessage
-            .inline_qos()
-            .parameter()
-            .iter()
-            .find(|&x| x.parameter_id() == PID_KEY_HASH)
-        {
-            Some(p) => {
-                if let Ok(key) = <[u8; 16]>::try_from(p.value()) {
-                    Some(key)
-                } else {
-                    None
-                }
-            }
-            None => None,
-        };
-
-        Ok(CacheChange {
-            kind,
-            writer_guid: Guid::new(source_guid_prefix, data_submessage.writer_id()).into(),
-            source_timestamp: source_timestamp.map(Into::into),
-            instance_handle,
-            sequence_number: data_submessage.writer_sn(),
-            data_value: data_submessage.serialized_payload().clone().into(),
-        })
-    }
-}
-
-pub struct RtpsReader {
-    endpoint: RtpsEndpoint,
-}
-
-impl RtpsReader {
-    pub fn new(endpoint: RtpsEndpoint) -> Self {
-        Self { endpoint }
-    }
-
-    pub fn guid(&self) -> Guid {
-        self.endpoint.guid()
-    }
-
-    pub fn unicast_locator_list(&self) -> &[Locator] {
-        self.endpoint.unicast_locator_list()
-    }
-
-    pub fn multicast_locator_list(&self) -> &[Locator] {
-        self.endpoint.multicast_locator_list()
-    }
-}
-
-pub struct RtpsStatelessReader {
-    guid: Guid,
-    topic_name: String,
-    history_cache: Box<dyn ReaderHistoryCache>,
-}
-
-impl RtpsStatelessReader {
-    pub fn new(guid: Guid, topic_name: String, history_cache: Box<dyn ReaderHistoryCache>) -> Self {
-        Self {
-            guid,
-            topic_name,
-            history_cache,
-        }
-    }
-
-    pub fn guid(&self) -> Guid {
-        self.guid
-    }
-
-    pub fn topic_name(&self) -> &str {
-        &self.topic_name
-    }
-
-    pub fn on_data_submessage_received(
-        &mut self,
-        data_submessage: &DataSubmessage,
-        source_guid_prefix: GuidPrefix,
-        source_timestamp: Option<messages::types::Time>,
-    ) {
-        if data_submessage.reader_id() == ENTITYID_UNKNOWN
-            || data_submessage.reader_id() == self.guid.entity_id()
-        {
-            if let Ok(change) = CacheChange::try_from_data_submessage(
-                data_submessage,
-                source_guid_prefix,
-                source_timestamp,
-            ) {
-                // Stateless reader behavior. We add the change if the data is correct. No error is printed
-                // because all readers would get changes marked with ENTITYID_UNKNOWN
-                self.history_cache.add_change(change);
-            } else {
-                error!("Error converting data submessage to reader cache change. Discarding data")
-            }
-        }
-    }
-}
 
 pub struct RtpsStatefulReader {
     guid: Guid,
